# Plotter and waterfall performance enhancements