… README.md
… src/qtgui/plotter.h