i run this application on my laptop and find that the CPU is quite high, so i did some flame graph analysis to see if i could reduce CPU (to have longer battery life). i had already run volk_profile. i found that actually the rendering was taking up more CPU than the actual GNUradio and DSP operations. i did the following operations:

- cache meter font so it can be reused
- update meter less often (only when the 0.1's place needed updating). the extra renders caused excessive blit operations when the CPlotter was drawing but the CMeter was updating the back buffer. they will still happen, but ideally significantly less often. a better solution could be to sync up the CMeter refreshes with the CPlotter, but i held off on that.
- cache peak circles as QImage so we dont need to calculate ellipse rasters on each frame for each ellipse occurence. the raster is the same and identical so a simple QImage draw will be faster.
- sliding window waterfall implementation. this eliminates a large memmove on each frame. instead the line is written across the top of a QImage which is 2x height in size and the "window" to be copied is slided up. once the window hits the top, then the top is copied to the bottom and the window sliding starts from the bottom back to the top. see diagram.
- use polygons for fills (instead of single line fills). each fill was attempted to be done in parallel. but the solution was to do fewer fills.
- prevent rounding errors. the plotter vs waterfall slider ratio was yielding fractional values sometimes. instead make sure the split is even. this make the drawing simple memmoves instead of a more expensive scale operation. i did this by flipping the math operations to ensure we are working with full pixel values before multiplying by the device pixel ratio.
- avoid unnecessary fills on mac: unifiedTitleAndToolBarOnMac makes the window translucent which means that the window must be cleared before each render. this is unnecessary since the painting is opaque.
- use solid colors. the alpha composition was more expensive. by precomputing the colors to be solid colors the composition is faster since we are copying and avoiding a bunch of extra math to do the alpha blending. i made the end result looks basically identical.
- store QPoint buffers on CPlotter object. allocate them once and reuse them.
- use opaque flag on widget. we are indeed painting every pixel so this flag should be set. disable autofill background.

sliding window implementation:
<img width="537" alt="image" src="https://github.com/user-attachments/assets/97c9b59e-26d9-46cd-b3c1-29ed9173a441">

i ran a test with these settings and docks open:
<img width="1680" alt="Screen Shot 2024-10-15 at 20 19 52" src="https://github.com/user-attachments/assets/abd0ccaf-73c6-4869-8418-9c3de6ef9c1a">


before:
![2024-10-15_20 19 01-WindowServer_sample txt](https://github.com/user-attachments/assets/5d98220b-75df-4ae6-a686-ce4cce1befa5)

after:
![2024-10-15_20 50 44-WindowServer_sample txt](https://github.com/user-attachments/assets/4e39b66c-95f4-4063-9183-a0a84f6e0916)

notable differences:
![savings](https://github.com/user-attachments/assets/dcc4a8bf-3799-46fd-b93d-972de9f89562)

i was doing 10fps at  110%, but now that is around 60%.
i was topping out at around 20 fps with around 130% cpu, but now that is around 100% cpu.

now i'm topping out around 40fps with around 110% cpu, so these changes provide a significant improvement.

open question: should the dock audio CPlotter share the same fps as the main CPlotter? it seems to be preset to 25fps (1000/40ms).