@yuzawa-san 2024-10-16
@willcode here are some screenshots. note the version in the window's title
<img width="1680" alt="Screen Shot 2024-10-16 at 08 49 41" src="https://github.com/user-attachments/assets/271b9a3e-8add-4f32-9847-0b9738258a5c">
<img width="1680" alt="Screen Shot 2024-10-16 at 08 50 23" src="https://github.com/user-attachments/assets/1845e195-ffbb-4df9-9b10-37c727afc0ba">