@argilo 2024-10-16
Peak circles appear to be drawn one or two pixels too far to the right.

Before:

![Screenshot from 2024-10-16 18-53-53](https://github.com/user-attachments/assets/10d732c8-1fb7-4387-bc7d-a866cb2ecace)

After:

![Screenshot from 2024-10-16 18-52-47](https://github.com/user-attachments/assets/b81a53e6-86f2-4c8e-8866-44691ca307eb)