@argilo 2024-10-17
> the line is written across the top of a QImage which is 2x height in size and the "window" to be copied is slided up. once the window hits the top, then the top is copied to the bottom

Is it necessary to use 2x height and copy? It might be possible to get the same performance benefit by splitting the `drawImage` call into two `drawImage` calls, one to draw from the "current" line to the bottom of the QImage, and a second to draw from the top of the QImage to the line above the "current" line. That would avoid the extra memory usage and eliminate the need for a copy.