@argilo 2024-10-17
The fill appears to be drawn one pixel lower than the line:

![Screenshot from 2024-10-16 20-40-10](https://github.com/user-attachments/assets/ce35e6c7-1404-49aa-bd8c-4efee4d12dd0)