@argilo 2024-10-18
Peak circles still seem to be drawn too far to the right:

![Screenshot from 2024-10-18 17-17-25](https://github.com/user-attachments/assets/a60b39e9-9f4a-4e1b-a2d2-a58bec5d7135)