@argilo 2024-10-18
If the Split Plot/WF slider is dragged all the way to the left (i.e. waterfall only), then the plot remains visible and quickly cycles through two or more sizes in an infinite loop.