@argilo 2024-10-18
> If the "Split Plot/WF" slider (in the "FFT Settings" panel) is dragged all the way to the left (i.e. waterfall only), then the plot remains visible and quickly cycles through two or more sizes in an infinite loop.

It looks like this bug happens because `plotWidthT` is zero in this case (because the `if (!m_2DPixmap.isNull())` block doesn't execute).