@yuzawa-san 2024-10-19
I have fixed both items. I actually was able to test in two device pixel ratios (1 and 2) with an external monitor. The DPR 1 was definitely broken as you reported. The DPR 2 I thought generally looked ok, but I was able to apply a fix to both of them. It had to do with some rounding and the fact that the the center of the circle is technically in the top left of the center pixel, so but nudging the pixmap over by one pixel we are back on alignment.

Here are my screenshots (note i enabled a background rect just for debugging, that has been removed):

DPR 1:
![Screen Shot 2024-10-18 at 20 56 15](https://github.com/user-attachments/assets/a29b0e89-9335-48e5-a0a3-5eaeee068ebd)

DPR 2:
<img width="1680" alt="Screen Shot 2024-10-18 at 20 56 41" src="https://github.com/user-attachments/assets/1237bbbf-2fa7-4ce7-8eb7-e4b39fed1108">