@argilo 2024-10-19
Thanks. It looks good now.

I'll get this merged soon, as long as no other bugs come up.