@argilo 2024-10-20
It appears this may have caused a performance regression: #1386

Any idea why that might be @yuzawa-san ?